+++ src/Recyclarr.Cli/Console/Settings/ISyncSettings.cs
@@ -1,4 +1,3 @@
-using System.IO.Abstractions;
 using Recyclarr.TrashLib.Config;
 
 namespace Recyclarr.Cli.Console.Settings;
@@ -6,7 +5,7 @@ namespace Recyclarr.Cli.Console.Settings;
 public interface ISyncSettings
 {
     SupportedServices? Service { get; }
-    IReadOnlyCollection<IFileInfo> Configs { get; }
+    IReadOnlyCollection<string> Configs { get; }
     bool Preview { get; }
     IReadOnlyCollection<string>? Instances { get; }
 }